+++ atdgen/test/test_ambiguous_variant_j.expected.ml
@@ -183,3 +183,20 @@ let read_ambiguous = (
 )
 let ambiguous_of_string s =
   read_ambiguous (Yojson.Safe.init_lexer ()) (Lexing.from_string s)
+
+
+(** {3 Generic Modules } *)
+module Ambiguous' = struct
+type nonrec t = ambiguous'
+let write = write_ambiguous'
+let read = read_ambiguous'
+let to_string = string_of_ambiguous'
+let of_string = ambiguous'_of_string
+end
+module Ambiguous = struct
+type nonrec t = ambiguous
+let write = write_ambiguous
+let read = read_ambiguous
+let to_string = string_of_ambiguous
+let of_string = ambiguous_of_string
+end