+++ CHANGES.md
@@ -1,3 +1,9 @@
+Unreleased
+----------
+
+* atdgen: Add option `-j-gen-modules` to generate JSON generic submodules (#420)
+
+
 2.16.0 (2025-01-22)
 -------------------